Move SpyWindowManager and WindowManagerCall out of the test file
