+++ tests/fakes.py
@@ -95,3 +95,50 @@ class FakeWindowManager(interfaces.TilingWindowManager):
 
     def get_window_sizes(self):
         return self._window_sizes
+
+
+class SpyWindowManager(interfaces.TilingWindowManager):
+    """Gets passed into LayoutManagers using dependency injection
+    and spies on their calls so we can make sure that we're handling
+    window creation correctly
+    """
+
+    def __init__(self, num_workspace_windows: int = 0):
+        self._calls: List[dtos.WindowManagerCall] = []
+        self._num_workspace_windows = num_workspace_windows
+
+    def make_window(
+        self,
+        window_details: dtos.WindowDetails,
+    ) -> None:
+        self._calls.append(dtos.WindowManagerCall(command="make", arg=window_details))
+
+    @property
+    def calls(self):
+        return self._calls
+
+    @property
+    def num_workspace_windows(self):
+        return self._num_workspace_windows
+
+    def resize_width(
+        self, target_window: dtos.WindowDetails, container_percentage: int
+    ) -> None:
+        pass
+
+    def resize_height(
+        self, target_window: dtos.WindowDetails, container_percentage: int
+    ) -> None:
+        pass
+
+    def focus(self, target_window: dtos.WindowDetails) -> None:
+        self._calls.append(dtos.WindowManagerCall("focus", arg=target_window))
+
+    def split(self, split_type: str) -> None:
+        self._calls.append(dtos.WindowManagerCall("split", arg=split_type))
+
+    def get_tree(self):
+        pass
+
+    def get_window_sizes(self) -> Dict:
+        pass
